requirements.txt
dash==2.17.0
dash-bootstrap-components==1.6.0
plotly==5.22.0
pandas==2.2.0
mysql-connector-python==8.2.0
python-dotenv==1.0.1
openpyxl==3.1.2
diskcache==5.6.3
